refatorando o componente UploadCSV
